+++ projetCrypto/Base_Donnee.h
@@ -5,15 +5,15 @@
 
 class Base_Donnee
 {
-	using int_trans = std::pair<int8_t, Transaction>;
+	using int_trans = std::pair<int, Transaction>;
 public:
 	Base_Donnee();
 	~Base_Donnee();
 
 	int_trans get(string)		 const;
-	int8_t    get_status(const string&) const;
-	void	  update(const Block&, int8_t);
-	void	  update(const string&, int8_t);
+	int    get_status(const string&) const;
+	void	  update(const Block&, int);
+	void	  update(const string&, int);
 	void	  print() const;
 	bool      push_back(const Transaction&);
